Feature: Stopword integration
